+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## V0.8.3
-- (WIP) Added: Created fix for the vanilla hand tracking lag while moving the player
+- Added: Fixed all player-relative VR elements lagging behind during movement and warp
 
 ## V0.8.2
 - Added: `jump` and `duck` (crouch) can now be bound in the Advanced Controls Manager.